requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI and ML
openai>=1.6.1,<2.0.0
langchain==0.0.350
langchain-openai==0.0.2

# Testing frameworks
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Web testing
selenium==4.15.2
webdriver-manager==4.0.1

# Mobile testing
Appium-Python-Client==3.1.1

# API testing
requests==2.31.0
httpx==0.25.2

# Data processing
pandas==2.1.4
numpy==1.25.2

# Utilities
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1

# Additional dependencies
beautifulsoup4==4.12.2
lxml==4.9.3
scikit-learn>=1.3.2
openpyxl>=3.1.0
xlsxwriter==3.1.9 